+++ contracts/0.8.9/interfaces/IStakingRouter.sol
@@ -11,7 +11,7 @@ interface IStakingRouter {
         view
         returns (address[] memory recipients, uint96[] memory stakingModuleFees, uint96 totalFeee, uint256 precisionPoints);
 
-    function deposit(uint256 maxDepositsCount, uint24 stakingModuleId, bytes calldata depositCalldata) external payable returns (uint256);
+    function deposit(uint256 maxDepositsCount, uint256 stakingModuleId, bytes calldata depositCalldata) external payable returns (uint256);
 
     /**
      * @notice Set credentials to withdraw ETH on ETH 2.0 side after the phase 2 is launched to `_withdrawalCredentials`
@@ -62,31 +62,31 @@ interface IStakingRouter {
         uint16 _treasuryFee
     ) external;
 
-    function updateStakingModule(uint24 _stakingModuleId, uint16 _targetShare, uint16 _stakingModuleFee, uint16 _treasuryFee) external;
+    function updateStakingModule(uint256 _stakingModuleId, uint16 _targetShare, uint16 _stakingModuleFee, uint16 _treasuryFee) external;
 
-    function getStakingModule(uint24 _stakingModuleId) external view returns (StakingModule memory);
+    function getStakingModule(uint256 _stakingModuleId) external view returns (StakingModule memory);
 
     function getStakingModulesCount() external view returns (uint256);
 
-    function getStakingModuleStatus(uint24 _stakingModuleId) external view returns (StakingModuleStatus);
+    function getStakingModuleStatus(uint256 _stakingModuleId) external view returns (StakingModuleStatus);
 
-    function setStakingModuleStatus(uint24 _stakingModuleId, StakingModuleStatus _status) external;
+    function setStakingModuleStatus(uint256 _stakingModuleId, StakingModuleStatus _status) external;
 
-    function pauseStakingModule(uint24 _stakingModuleId) external;
+    function pauseStakingModule(uint256 _stakingModuleId) external;
 
-    function resumeStakingModule(uint24 _stakingModuleId) external;
+    function resumeStakingModule(uint256 _stakingModuleId) external;
 
-    function getStakingModuleIsStopped(uint24 _stakingModuleId) external view returns (bool);
+    function getStakingModuleIsStopped(uint256 _stakingModuleId) external view returns (bool);
 
-    function getStakingModuleIsDepositsPaused(uint24 _stakingModuleId) external view returns (bool);
+    function getStakingModuleIsDepositsPaused(uint256 _stakingModuleId) external view returns (bool);
 
-    function getStakingModuleIsActive(uint24 _stakingModuleId) external view returns (bool);
+    function getStakingModuleIsActive(uint256 _stakingModuleId) external view returns (bool);
 
-    function getStakingModuleKeysOpIndex(uint24 _stakingModuleId) external view returns (uint256);
+    function getStakingModuleKeysOpIndex(uint256 _stakingModuleId) external view returns (uint256);
 
-    function getStakingModuleLastDepositBlock(uint24 _stakingModuleId) external view returns (uint256);
+    function getStakingModuleLastDepositBlock(uint256 _stakingModuleId) external view returns (uint256);
 
-    function getStakingModuleActiveKeysCount(uint24 _stakingModuleId) external view returns (uint256);
+    function getStakingModuleActiveKeysCount(uint256 _stakingModuleId) external view returns (uint256);
 
     function getKeysAllocation(uint256 _keysToAllocate) external view returns (uint256 allocated, uint256[] memory allocations);
